[WIP] Basic functionality added
Paths for files are static.
Creation of new Builds.
Reading of skills config files.
Cap level of Skills will decrease as the level of the char decreases.
Skill deactivation when level is lower then the minimum level for the skill.
Prerequirements are set as a skill with prrequirements is leveled up.
[might need some tweeking]
Points Total updates as the skill is leveled.
Saving of the Builds. [tweeking required]
Namechanges in List, when the Name of the Build is changed.
Custom pictures possible when creating new build

TODO:
Prerequirements check
copy/pasting builds.
saving of the custom pictures
custom saving path
loading
editing of pictures
QP-skills
Quests

creating all the config files :/
reading paths from config

things I can't think of right now